+++ packages/vsce/__tests__/__e2e__/util/constants.ts
@@ -64,4 +64,6 @@ export const PLIB2NONE = "PLIB2NONE";
 export const CBSA = "CBSA";
 
 //JVMServer labels
-export const EYUCMCIJ = "EYUCMCIJ";
+export const JAHAGWLP = "JAHAGWLP";
+export const JVMEWLP = "JVMEWLP";
+export const JVMDIWLP = "JVMDIWLP (Disabled)";