+++ packages/vsce/src/doc/meta/localFile.meta.ts
@@ -11,11 +11,9 @@
 
 import { CicsCmciConstants } from "@zowe/cics-for-zowe-sdk";
 import { Resource } from "../../resources/Resource";
-import { PersistentStorage } from "../../utils/PersistentStorage";
+import PersistentStorage from "../../utils/PersistentStorage";
 import { ILocalFile } from "../resources";
 import { IResourceMeta } from "./IResourceMeta";
-
-const persistentStorage = new PersistentStorage("zowe.cics.persistent");
 
 export const LocalFileMeta: IResourceMeta<ILocalFile> = {
   resourceName: CicsCmciConstants.CICS_CMCI_LOCAL_FILE,
@@ -26,8 +24,8 @@ export const LocalFileMeta: IResourceMeta<ILocalFile> = {
     return criteria.map((n) => `file=${n}`).join(" OR ");
   },
 
-  async getDefaultCriteria () {
-    return PersistentStorage.getDefaultFilter(CicsCmciConstants.CICS_CMCI_LOCAL_FILE, "localFile");
+  getDefaultCriteria() {
+    return PersistentStorage.getDefaultResourceFilter(CicsCmciConstants.CICS_CMCI_LOCAL_FILE, "localFile");
   },
 
   getLabel: function (localFile: Resource<ILocalFile>): string {
@@ -47,9 +45,8 @@ export const LocalFileMeta: IResourceMeta<ILocalFile> = {
   },
 
   getContext: function (localFile: Resource<ILocalFile>): string {
-    return `${
-      CicsCmciConstants.CICS_CMCI_LOCAL_FILE
-    }.${localFile.attributes.enablestatus.toUpperCase()}.${localFile.attributes.openstatus.toUpperCase()}.${localFile.attributes.file}`;
+    return `${CicsCmciConstants.CICS_CMCI_LOCAL_FILE
+      }.${localFile.attributes.enablestatus.toUpperCase()}.${localFile.attributes.openstatus.toUpperCase()}.${localFile.attributes.file}`;
   },
 
   getIconName: function (localFile: Resource<ILocalFile>): string {
@@ -101,10 +98,10 @@ export const LocalFileMeta: IResourceMeta<ILocalFile> = {
   },
 
   async appendCriteriaHistory(criteria: string) {
-    await persistentStorage.addLocalFileSearchHistory(criteria);
+    await PersistentStorage.appendSearchHistory(CicsCmciConstants.CICS_CMCI_LOCAL_FILE, criteria);
   },
 
   getCriteriaHistory() {
-    return persistentStorage.getLocalFileSearchHistory();
+    return PersistentStorage.getSearchHistory(CicsCmciConstants.CICS_CMCI_LOCAL_FILE);
   },
 };