+++ packages/vsce/CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to the "cics-extension-for-zowe" extension will be documente
 
 - Validate input using maximum resource length when inspecting resource. [#402](https://github.com/zowe/cics-for-zowe-client/issues/402)
 - Action menus on Resource Inspector Header. [#333](https://github.com/zowe/cics-for-zowe-client/issues/333)
+- Enhancement: Added Enable/Disable actions to JVM Servers [#391](https://github.com/zowe/cics-for-zowe-client/issues/391)
 
 ## `3.11.0`